bug(cards): add border to the cars
[Finishes #168250575]
