+++ UI/css/style.css
@@ -515,10 +515,20 @@ input[type="submit"] {
     max-width: 960px;
     max-width: 90%;
     margin: 0 auto 30px;
+    
+   
   }
   
   article {
     position: relative;
+    border-left: 0.09px solid #233245;
+    border-right: 0.1px solid #233245;
+    border-bottom: 0.1px solid #f1f1f1;
+    /* border: 1px solid #2196F3; */
+    background-color: #f0f0f1;
+  }
+  .modelbtn1{
+    border-right: 0.2px solid #233245;
   }
   
   .article-img {
@@ -574,7 +584,8 @@ input[type="submit"] {
     background-color: #fff;
     margin: auto;
     /* width: 50%; */
-    width: 80%; 
+    width: 80%;
+ 
     border-radius: 12px;
     
     animation-name: modelAnimation;
@@ -617,14 +628,14 @@ input[type="submit"] {
 }
 
 .main-image img{
-width:98%;
+width:95%;
 height:100%;
 border-radius: 2%;
 
 }
 
 .details{
-display:flex; flex-direction:column;justify-content: space-between;
+display:flex; flex-direction:column;justify-content: space-between; margin-top: 15px; margin-bottom: 15px;
 }
 @media screen and (max-width: 790px) {
     .modalLayout-content {
