+++ UI/html/View Mentors.html
@@ -70,7 +70,7 @@
               <p><b>Name:</b> Amani Murera</p>
               <P><b> Senior Software Developer</b></P>
           </div>
-          <button  class="button" >Read More</button>
+          <button  class="button" style="border-radius:0px" >Read More</button>
         </article>
         <article class="modelbtn">
           <img class="article-img" src="https://www.mantelligence.com/wp-content/uploads/2017/10/Questions-To-Ask-Smile-or-eyes.jpg" alt="" />
@@ -227,21 +227,22 @@
                 <p> <b>Location:</b> <span>Kigali</span> </p>
                 <p> <b>Company:</b> <span>Andela</span> </p>
                 <p> <b>Position:</b> <span>Software Developer</span> </p>
-                <p> <b>Area Of Expertise:</b> <span>Web Technology</span> </p>
-                <p> <b>Experience:</b> <span>1 Year </span> </p>
+            
+                <!-- <p> <b>Area Of Expertise:</b> <span>Web Technology</span> </p> -->
+                <!-- <p> <b>Experience:</b> <span>1 Year </span> </p> -->
     
     
     
-                <article id="description">
+                <section id="description">
     
-                    <h3> Description </h3>
+                    <h3> Description: </h3>
     
                     <p> Lorem ipsum, dolor sit amet consectetur adipisicing elit. Accusamus reprehenderit, consectetur, rem
                         laudantium tempore cum perspiciatis corrupti, saepe itaque culpa at illum iste exercitationem
                         numquam
                         voluptatibus eius molestiae eos cupiditate.</p>
-                </article>
-    
+                    </section>
+              <!-- </br> -->
                 <div id="buttons">
                     <a href="Request  Mentor.html"><button class="modelbtn button">Create Request Session</button></a>
                 </div>
