+++ server/helpers/tokenGenerator.js
@@ -1,7 +1,7 @@
 import jwt from 'jsonwebtoken';
 
 const generateAuthToken = (id, admin, mentor) => {
-  const token = jwt.sign({ id, is_admin: admin, is_mentor: mentor}, process.env.Token_Key);
+  const token = jwt.sign({ id, is_admin: admin, is_mentor: mentor }, process.env.Token_Key);
   return token;
 };
