+++ server/models/session_model.js
@@ -77,7 +77,7 @@ class Session {
   // reject sessions
   reject = (res, id) => {
     const session = this.sessions.find((sid) => sid.sessionId === parseInt(id, 10));
-    if (!session) return res.status(404).send({ status: 404, error: 'this session  is not found!' });
+    if (!session) return res.status(status.NOT_FOUND).send({ status: status.NOT_FOUND, error: 'this session  is not found!' });
     if (session.status === 'Accept') {
       return res.status(status.FORBIDDEN).send({
         status: status.FORBIDDEN,
